+++ Source/Lib/Common/Codec/filter.h
@@ -32,8 +32,6 @@ extern "C" {
 #define SCALE_EXTRA_BITS (SCALE_SUBPEL_BITS - SUBPEL_BITS)
 #define SCALE_EXTRA_OFF ((1 << SCALE_EXTRA_BITS) / 2)
 
-typedef int16_t InterpKernel[SUBPEL_TAPS];
-
 #define BIL_SUBPEL_BITS 3
 #define BIL_SUBPEL_SHIFTS (1 << BIL_SUBPEL_BITS)